pyproject.toml
[project]
name = "car-accident-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.8.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
]
